Open sqlite3 connections using uri mode
This appears to be the only way to pass additional options to sqlite3,
including the ?cache=shared parameter.

The previous changeset to enable cache sharing was broken, but appeared to work
because sqlite3 interpreted ':memory:?cache=shared' as a file name, so the test
suite was running against a file-backed database.