tests: refactor test fixtures to use `backend` pytest fixture
Using the `backend` pytest fixture instead of calling `get_backend`
removes the need to add cleanup code in each and every test. This has become
necessary because sqlite3 connections have changed to use a shared cache,
and so if the db is not cleaned up between tests we get spurious failures.